# swagger.yaml
openapi: "3.0.2"
info:
  title: "Puppy Shop"
  version: "1.0"
servers:
  - url: http://localhost:10778/
    description: shop server
paths:
...
  /products:
    get:
      description: get all products
      security:
        - token: []
          refreshToken: []
      responses:
        "200":
          $ref: "#/components/responses/products"
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      description: add product - admin only
      security:
        - token: []
          refreshToken: []
      responses:
        "200":
          $ref: "#/components/responses/products"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "401":
          $ref: "#/components/responses/unauthorized"
  /products/{id}:
    put:
      description: edit product -admin only
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: object
            properties:
              id:
                type: integer
                example: 4
      security:
        - token: []
          refreshToken: []
      responses:
        "200":
          $ref: "#/components/responses/products"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "401":
          $ref: "#/components/responses/unauthorized"
  /products/count:
    get:
      description: get amount of products in site for welcome page
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  err:
                    type: boolean
                    example: false
                  productsCount:
                    type: integer
                    example: 20
        "500":
          $ref: "#/components/responses/InternalServerError"
  /products/category:
    get:
      description: get all product categories
      security:
        - token: []
          refreshToken: []
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  err:
                    type: boolean
                    example: false
                  categories:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          example: 2
                        name:
                          type: string
                          example: Food
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /orders:
    get:
      description: get last order for user
      security:
        - token: []
          refreshToken: []
      responses:
        "200":
          $ref: "#/components/responses/order"
        "204":
          description: data not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  err:
                    type: boolean
                    example: true
                  msg:
                    type: string
                    example: There is no last order
        "500":
          $ref: "#/components/responses/InternalServerError"
        "401":
          $ref: "#/components/responses/unauthorized"
    post:
      description: add new order
      security:
        - token: []
          refreshToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                order_total_price:
                  type: integer
                  example: 200
                cart_id:
                  type: integer
                  example: 3
                user_id:
                  type: integer
                  example: 6
                city:
                  type: string
                  example: Ashdod
                street:
                  type: string
                  example: Katz
                delivery_date:
                  type: string
                  example: "2021-05-05"
                closing_date:
                  type: string
                  example: "2021-04-05 17:45:30"
                credit_card:
                  type: integer
                  example: 1234
      responses:
        "200":
          $ref: "#/components/responses/order"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "401":
          $ref: "#/components/responses/unauthorized"
  /orders/count:
    get:
      description: count number of orders in site
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  err:
                    type: boolean
                    example: false
                  numberOfOrders:
                    type: integer
                    example: 48
        "500":
          $ref: "#/components/responses/InternalServerError"
        "401":
          $ref: "#/components/responses/unauthorized"
  /orders/dates:
    get:
      description:  get orders dates for filtering the delivery day in order form
      security:
        - token: []
          refreshToken: []
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  err:
                    type: boolean
                    example: false
                  filteredDates:
                    type: array
                    items:
                      type: object
                      properties:
                        delivery_date:
                          type: string
                          example: "2021-05-14"
                        counter:
                          type: string
                          example: 3
        "500":
          $ref: "#/components/responses/InternalServerError"
        "401":
          $ref: "#/components/responses/unauthorized"
  /orders/{order_id}/receipt:
    get:
      description: get receipt file
      parameters:
        - in: path
          name: order_id
          required: true
          schema:
            type: object
            properties:
              id:
                description: order id
                type: integer
                example: 4
      security:
        - token: []
          refreshToken: []
      responses:
        "200":
          description: The receipt of the current order
        "400":
          description: bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  err:
                    type: boolean
                    example: false
                  msg:
                    type: string
                    example: error
        "404":
          description: data not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  err:
                    type: boolean
                    example: true
                  msg:
                    type: string
                    example: There is no order with this id
        "500":
          $ref: "#/components/responses/InternalServerError"
        "401":
          $ref: "#/components/responses/unauthorized"
components:
  securitySchemes:
    token:
      type: apiKey
      in: header
      name: token
    refreshToken:
      type: apiKey
      in: header
      name: refreshToken
  responses:
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            type: object
            properties:
              err:
                type: boolean
                example: true
              msg:
                type: string
                example: error
    unauthorized:
      description: unauthorized action
      content:
        application/json:
          schema:
            type: object
            properties:
              err:
                type: boolean
                example: true
              msg:
                type: string
                example: unauthorized action
    products:
      description: products
      content:
        application/json:
          schema:
            type: object
            properties:
              err:
                type: boolean
                example: false
              products:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                      example: 1
                    name:
                      type: string
                      example: product1
                    category_id:
                      type: integer
                      example: 2
                    price:
                      type: integer
                      example: 119
                    image:
                      type: string
                      example: url.string
                    product_amount:
                      type: integer
                      example: 3
...
    order:
      description: last/new order
      content:
        application/json:
          schema:
            type: object
            properties:
              err:
                type: boolean
                example: false
              order:
                type: object
                properties:
                    id: 
                      type: integer
                      example: 1
                    user_id:
                      type: integer
                      example: 3
                    cart_id: 
                      type: integer
                      example: 4
                    order_total_price: 
                      type: integer
                      example: 213
                    city: 
                      type: string
                      example: Ashdod
                    street:
                      type: string
                      example: Katz
                    delivery_date: 
                      type: string
                      example: "2021-05-05"
                    closing_date: 
                      type: string
                      example: "2021-04-05 17:45:30"
                    credit_card: 
                      type: integer
                      example: 1234